警告つぶし。

+++ NVEncCore/rgy_output_avcodec.cpp
@@ -2270,7 +2270,7 @@ RGY_ERR RGYOutputAvcodec::WriteNextFrameInternal(RGYBitstream *bitstream, int64_
         const auto frameI = (frameType & (RGY_FRAMETYPE_IDR | RGY_FRAMETYPE_I)) != 0;
         auto& qVideoQueueFree = (frameI) ? m_Mux.thread.qVideobitstreamFreeI : m_Mux.thread.qVideobitstreamFreePB;
         auto queueFavoredSize = (frameI) ? VID_BITSTREAM_QUEUE_SIZE_I : VID_BITSTREAM_QUEUE_SIZE_PB;
-        if (qVideoQueueFree.size() > queueFavoredSize) {
+        if ((int64_t)qVideoQueueFree.size() > queueFavoredSize) {
             //あまり多すぎると無駄にメモリを使用するので減らす
             bitstream->clear();
         } else {
